Changed the way the User class works
User class now takes care of checking for correct input, throwing an InputMismatchException in its constructor if things are wrong.

Genders is no longer a public enum, it is local to the User class. Instead the User class takes gender as a String and then personally makes it into the appropriate enum value.